:recycle: separate file for post diary store
